qcore: make DisallowInheritance work with six.with_metaclass
Seems to be some bad interaction with Cython.
