+++ qcore/tests/test_disallow_inheritance.py
@@ -1,0 +1,13 @@
+import qcore
+from qcore.asserts import AssertRaises
+import six
+
+
+class Foo(six.with_metaclass(qcore.DisallowInheritance)):
+    pass
+
+
+def test_disallow_inheritance():
+    with AssertRaises(TypeError):
+        class Bar(Foo):
+            pass
